--- src/services/cart-services/getCartService.js
@@ -1,9 +1,0 @@
-import axios from "axios";
-
-const getCartService = async( encodedToken ) => await axios.get('/api/user/cart',{
-    headers: {
-        authorization: encodedToken
-    },
-})
-
-export default getCartService;
